Fix Issue 874
